+++ src/pages/schemas/joi.update-page.schema.ts
@@ -16,6 +16,7 @@ export const updatePageSchema = Joi.object({
     .valid(...Object.values(PageStatus))
     .optional(),
   isHomepage: Joi.boolean().optional(),
+  isTitleVisible: Joi.boolean().optional(),
   seo: seoSchema.optional(),
   blocks: pageBlocksSchema.optional(),
 });